Add skeleton for states, drop fb-functions
As I want to run many of the logic on the client side to avoid load
on the server, where possible, And also as I want the logic to be
visible to the end user, so dropping the bringing in of fb-functions.

Have added some skeletons towards showing states available in the
database (inturn hospitals available) at a later stage.

Drop the features check and listing on the page.