MoveUp elStates, Update its text with State Names
Make elStates global to the script. So that it can be accessed
from any function in there.

Add logic to update the elStates textContent with each got states'
name.

Add a margin to a element style so that if two of them are placed
next to one another, there will be spacing between them, so that
it will appear neater.