FAdmin: Use named function with bind to identify which call
By default the promise returned by collection.set doesnt specify
which document it was be it wrt success (then) or chained failure
(catch),

SO I am using a named/normal function with bind to attach/prepend
additional info, rather in this case the state which is being setup.

NOTE: Even thou the promise is specified as working with a void,
on success it was found that some sort of time data is passed wrt
the then's resolve/success.
NOTE: Catch gets a error object, from which we extract the detail.

NOTE: When binding a function with additional arguments/data, pass
1st arg as null, whether the location which calls the function passes
a argument or not.