Allow listing max 25 hospitals in one shot
This should be fine, because, this app is mainly to find hospitals
with free beds, so finding max 25 hospitals in a district with free
beds should be fine.

NOTE: This also means one can't just get the list of all hospitals
and then select a hospital to update its data. Rather admins should
be directly taken to a page where they can update the hospitals they
are authorised for.