WebApp: GetStates doesnt manipulate UI
Now get_states just returns a list of states with their IDs, inturn
a seperate ui function takes care of creating the anchors within
the specified element.

As getting the docs in a collection is a asyncronous logic, and
we want the list of items from this, so we wait on this operation.
Inturn to await on this, we have to declare get_states itself as
a async function, which doesnt return a value, but instead returns
a promise, which has to be resolved. This is handled by the main
logic, when it calls the ui update logic with in the then of the
promise returned.