WebApp:UI: Fix table creation logic, create html only then set
If html table is built part by part by directly setting into
a html elements inner html property, it doesnt seem to work, bcas
chances are the concatenation logic is checking for validity of
the html and thus inturn clearing off the part table tags.

So now the full table html is built in a independent variable,
and then assigned to the innerHTML property at the end.