WebApp:DB: Return a sorted districts list wrt state
As the java object created to represent the document, didnt return
its member in a sorted order, so added explict sorting.

With this there should be consistance in UI, when users go about
back and forth and or revisit the same page again. Also from user
perspective now the districts will be shown in a sorted order,
which should make it easy for people to identify the district
they want from the big list more easily.