FirestoreRule: Pick Doc/HospId from request.resource
As the resource object corresponds to the end/new doc and
request.resource object corresponds to what the user has sent,
so if the request.resource is having a invalid resource id, then
resource object will be null. So one uses it to check validity
of request, the request will fail due to non existent end
resource rather than a explicit check. So changing to resource id
in the request and then checking for auth in my auth database.

Logically end result should be same, while at same time it should
allow the rule checks to go through fully and then fail.