WebApp: Update button handler, Get and Got Pass states
Added a skeletal handler for the update button, which inturn
sets the get pass flag and calls ui_sync, which takes care of
doing the required thing to put ui in the needed state.

Added a additional variable gbGotPass so that we can distinguish
between initially wanting to get a pass and then once pass as
been got.