WebApp: Put Auth UI in Auth Div, Named Auth Handler
As elMain sets html tags up to a custom visual look, so have changed
to a seperate elAuth div for the auth ui. Otherwise the buttons and
general look and feel of AuthUI was messing up.

Also moved the auth handling into a named func, which is created in
the ui file, and inturn it sets up ui for the next phase.