WebApp: gGotAuth, gbUpdate
Morphed gbGotAuth to gGotAuth, which inturn stores the UID fo the
authenticated user. Logics updated to take care of same.

Added a gbUpdate which will be set if update btn is pressed and
the user is authenticated. In this case instead of showing the
auth ui, a update screen needs to be shown.

	THis will involve getting the hospitals for which the
	user is the admin and show their details so that the
	user can update them.

The sign in success handler now also chains in the ui_sync func,
so that appropriate update ui can be shown.