WebApp: GetAdminHosps
Renamed the previous logic to get_adminhospitalids, as it retreives
only the relavent hospital ids.

Inturn a new function added, which retreives details about the matching
hospitals also. For this the where 'in' query mechanism of firestore
is used, this inturn limits the number of documents that can be
retrieved to 10 as it stands now, on top even I have put a duplicate
limit 10 condition.

Also it uses the FieldPath.documentIds() mechanism to filter based on
the document id rather than a field in the document.

The UI logic also updated to show a table, for the corresponding
hospitals, but unlike a normal mode UI, this update mode ui doesnt
show the timestamp. TODO: Later have to convert the Beds fields
to input/edit boxes. Also need to add individual update buttons,
next to each hospital entry in this case involving updating mode.