WebApp:UI:Hide home button
history.deleteAll related to browser seems to be different from
history actions related to pop and back and ..., so either need
to loop through backs or ... for going to home from a history
perspective, while also clearing all the nav till then. Need
to think through and look at this later. For now /dev/null.