WebApp: Add google site verification tag
When I entered the dummy test email which I was using for testing
in multiple devices and browsers, and in both online and offline
version of the site, google chrome poped a alert telling may be
password is compromised on a deceptive site, so just to hopefully
let google mama know that thats a wrong thing tried adding this.

Another possibility is, as the password is a simple one, maybe
someone else somewhere else also has same password or so, or it is
a relatively common password, so its flagging, not sure.