WebApp:DB: Load StatesUTsDists json, avoid Nw/FS queries for same
This should help avoid querying firestore wrt list of states and
wrt list of districts within the states.

As querying firestore costs money and network egress+costs and a
user may trigger it multiple times, also as I am currently not
sure of how firestore sdk caches things or not, so I am explicitly
downloading the StatesUTsDists json at the begining and then using
the data from local in memory copy.

This should also speed things up a bit wrt those phases/aspects,
beyond a initial few kbs of addtional nw load per session.

The admin script also works from a duplicate of the same json,
so both should be in sync. As long as the duplicates are in sync.