JS:Switching from CommonJS to ESMA modules and related conventions
Have changed from using require to import (as part of the move
to using ecma conventions and mechanisms).

Inturn one can import js files but not json files directly. So
have created a new statesuts_districts.mjs file which declares
a goStates object variable which is assigned from the json structure,
inturn the same is noted as exported in the mjs file.

NOTE: Even require would have created a object when importing json.
NOTE: One requires to name the js file as mjs to let node explicitly
and unambiguously know that we are using ecma module standards.
NOTE: Inturn have explicitly asked for the variable in the module
to be imported into the main scripts context.

Have noted that the top level index.js file follows the ecma conventions
by specifying the type as module in package.json. THis avoids the
need to explicitly rename the top level js file as mjs.

NOTE: A file with extension cjs is assumed to follow the commonjs
related conventions.