Revert "JS:Switching from CommonJS to ESMA modules and related conventions"
This reverts commit 50a8015c0882c00fa40f5f2e2983cd32699a3dbe.

The idea was to have a common mechanism/conventions across both the
WebAPP frontend and the backend scripts. But looking into the
documents it appears like the new ecma conventions based modules
support is still in beta only and also wrt admin sdks the picture
doesnt seem clear, as to where it stands wrt the transition fully.

So for now am reverting the changes done to admin script, so that
it goes back to following the CommonJS modules and related convention.