Firebase:Firestore:JS: the elusive error
Seems like the error object in some cases as the details and in
other doesnt. Same for even code. However message seems to put
something out there when the error occurs, and details seems to
put a part of message out there. Also need to check what is this
metadata and note that has tumbled out.