WebApp:DB+UI: Move UI status update to UI file
Instead of hard coding the UI related updating of a db operations
status in the db related file, now move it to be in the ui related
file itself, by using a callback mechanism with a custom user settable
argument.

Inturn use this new mechanism to change the color of the sync btn,
based on whether it was a success or not.