WebApp:UI:ui_select, ui_table with overwrite control
Added ui_select to allow drop down selection html widget to be
added into a html page.

The same is used to added a bedType selection for hospitals in
a district page. The handler needs to be implemented yet.

Inturn to allow both a ui_select and ui_table to coexist, the
ui_table logic has been updated to allow its user to control,
whether it will append to the html element passed to it or if
it will overwrite it.

Also as updating the elMain with ui_table after ui_select has
written to it, seems to clear the element handle associated
with ui_select's select element and inturn its event handlers,
so added a new ui_select_changehandler to allow the change
handler to be set again.

So if one is creating a ui_select on its own in a div element,
then passing change handler through ui_select is good enough.
However if one is sharing the div element between ui_select
and say ui_table or any other such logics, then better to
assign the change handler at the end using ui_select_changehandler.