WebApp: User selects Hospital paramater for hosp listing
Added support to db_get_hospitals to allow the hospital param/field
on which hospitals are selected and ordered. Also changed the minFree
from 5 to 1.

ui_select takes a sDefault argument, which sets the option/item
selected by default.

Inturn the above are used to allow the user to select as to based
on which parameter they want to filter the list of hospitals within
the district.