WebApp:DB: Fix ordering issue
Was using a firestore syntax from python in javascript, and as
javascript doesnt seem to rise a error if a named argument is passed
even if such a named argument is not there (need to check this aspect
once again), so I missed it. But while testing found that the data was
ording in ascending order, even thou I wanted descending order, so on
checking at the api reference realised the issue.