System: Ventilators status in db, Order of displayed fields
Have added a field to store ventilators status wrt each hospital.
Inturn logic udpated to handle the same be it wrt creating data
schema, sample data, displaying data or updating data.

The order of data fields is displayed is changed so that the
beds status is shown first before Name and pin code. 1 because
beds status is important and also because Name could be small or
large and thus can push the bed status info outside the screen
with mobiles in portrait mode or so. This shouldnt occur now.

Also renamed from bedType tag to hospParam, wrt ui code, because in
future, if required one may want to add oxygen availability etc.