Firestore indexes and rules updated
New indexes added to take care of descending order based listing
wrt all the currently supported 3 hosp paramaters ie icu, normal,
ventilator.

Added BedsVntltr to the list of fields allowed to be updated from
web/mobile apps.

NOTE: If a rule doesnt allow a particular field to be updated, and
inturn if a update request is sent with data for such a field, but
without changing its value, the update request will succeed and
update other fields which may have been updated and which are also
allowed to be updated. However if such a field's data is diff in
the update request, then it will be rejected.