FAdmin:DBX: Import Collection
Added a generic import collection routine and linked it to a command
usable by end user/admin.

This can be used to fill the system's database with any number of
collections of documents. Rather in this case it can help with
importing of actual hospitals, or hospital admin info.

The same can also be used to update a collection with new documents
and or to update contents of existing documents.