WebApp:AvailabilityStatusQuery Loading data message
If there is a disruption in the internet connection, this message
will be visible so that the user can get a rough idea of what might
have happened and what they may want to do.

If the connection comes back, the logic will update the screen with
needed data automatically in most cases.

NOTE: During normal operation this flashes on the screen, and appears
bit odd, but feel its useful for the worst case, for now.