FAdmin:HAdmin: getUID or randomBytes, connect to Auth Emulator
Updated setup-env.sh so that the admin script can connect to the
auth emulator during development and testing.

getUserId logic updated to return uid if the email passed is a
verified email. Else it returns a invalid uid having invalid
tag followed by random bytes.

NOTE: To test verified email with auth emulator, load the
verification url shown in the emulator logs into a browser.

Also to keep the logic simple for now have changed the import
function into async function, and inturn await for getUserId
function to return before proceeding to next admin's uid. This
should be fine, as async processing of the admin records during
admin importing is fine, as import function will handle it fully,
and users of hospadmins.import function dont have to worry about
what it returns, because it doesnt return anything to its caller
on its own. Tocheck later, if a dummy promise is still returned
for async function, which is possible, to may be help with
then.catch chaining or so, need to look at the docs. But in this
specific case it doesnt matter, in general.