FAdmin:HAdmin: Add to HospitalsExtra collection
If getUserId doesnt raise a exception then write the uid as adminid
for the current hospital. However do note that for non verified
email id's, it will be writing a randomly generated invalid uid,
along with a log to the console stderr.

So by either looking to stderr or by looking into the invalid uids
in the AdminId field, the system admin will have to inform the
dataowner/hosp admin involved to get their email verified with the
auth system (by clicking on the link sent to their email id).

And once the hosp admin get their email verified, the sys admin
needs to rerun the import_hospadmins command for such admins.