MakeDuplicatable:P1: Remove hardcoded databaseURL from admin script
The service account token which is specified using env variable will
contain the project details, so this can be avoided.

This ensures that, if someone wants to duplicate this system,
there is no need to modify the source file, instead just pointing
the env variable to their projects service account token will
do.