WebApp: Hide HospId in normal mode; to help fit mobile screens
As HospId is not useful for general users, so the same is hidden
in normal mode. It will be shown only in UpdateMode.

This is done so that in mobile portrait mode, even the hospital
name (or atleast a part of it, if it is very long) can appear
on the screen without needing to scroll horizontally.

With the reduction in min width of Bed availability fields and
this hiding of hosp id, hopefully it should be more useful by
default for users on mobile.

This is achieved by adding a hide type to mTypes passed to ui_table.