FAdmin: optional TEST mode wrt import_hospitals
Instead of making generate_statesuts_hosps generate wrong data
(i.e data with beds set to 1 instead of -1). NOW rather genhosps
is set to generate with beds value of -1 only.

And inturn import_hospitals command as a additional argument
that can be passed to tell whether we are importing in TEST
mode or normal mode. In TEST mode it will forcibly set beds
available values to 1 instead of what is there in the json file.

Thus while importing one can decide whether to import with
actual data available (or -1 in case there is no valid data,
which is the default for gen_hosps) or with a dummy 1 for now.