Scripts:Generate Regions Avoid my json creation logic
Now the json module is used to create the json file of the regions
rather than my handrolled json file creation logic.

Inturn the json created using this new logic is imported into
git.

NOTE: THis was done to simplify the logic or rather make the
actual logic clearer for anyone looking at it in future. ANd keep
the json structure related things in the standard json module.