Scripts:GenHosps: Include all hospitals by default
The meta data available from data.gov.in wrt hospitals seems to be
of mixed quality.

In some places govt hospitals are categorised accordingly and in
other places govt hospitals are not categorised as such at all.

The minimal categorisation info is inturn available only for around
10% or less of the hospitals in the dataset. And that to mainly wrt
hospitals in large cities and not at all for most other districts.

So to avoid filtering hospitals wrongly, am going back to including
all hospitals in the generated json file.