FAdmin: Add (Enable) Test Admin
Added a new command to set email verified flag for a given email
id, which is set to test@india.world by default. (Currently user
is not allowed to change it).

Need2Think whether to allow adding of user/admin email id into auth
system through this direct mechanism or not.