WebApp:OnDemandAuthUI: Chain each step
Done so that any delay in loading the required stylesheets doesnt
cause a UI redraw, well thats the idea.

NOTE: It seemed like with the on demand authui loading logic, the
auth ui was redrawing, which wasnt occuring with the older load
all things required at the begining itself. So hopefully this
chaining to ensure things are loaded dynamically as required b4
showing the auth ui, cleans this up.