FireStoreRules: Initial skeleton for Patients collection
As the rules dont seem to allow partial matching wrt collection/
document names, need to look at docs. So am thinking of a new data
schema wrt Patients. However need to check if a rule for a subcollection
can access the fields of the doc containing the subcollection in a
simple straight manner or not.

This also allows the list of alloted patients to be maintained within
the same hierarchy.