WebApp+: Fix a oversight wrt ToAllot SubCollection name
As part of realising this oversight, had to enable logs, as well
as allow access to the documents within Patients at the top level,
b4 I realised this issue.

TODO: Also realised that as I was not explicitly specifying var
for the variables within admin commands wrt creating testdata
(I havent done it at other places also, didnt realise that it
leads to global variables by default rather than local in js;
so need to review all the code in general wrt this), so if the
creation of the records is delayed by the underlying system,
then it would pickup the wrong/current_at_that_time value. I
noticed the 2nd state's 1st district getting assigned hospital
825 or so, because of this.

Also as I am currently running both the async funcs (i.e wrt
hospitals and patients) parallely, so need to think through
implications from this and globals and ...