JS:FAdmin: Update hospitals script to use var/let as required
In general have specified the variables in the function as var,
so that they have function scope. In turn for variables which
are used within a loop and are also passed to functions which
will be run asyncronously at a later time, have used let to
define a block scoped variable, which defines a new instance
each time through the loop (need to think through and check
this once).

Does JS ensure that the value is maintained independently till
the asyncronous calling of the callback/then/...

Also my use of bind should avoid any issue wrt the variables and
their use at a later time, ie for variables bound using bind. This
was done specifically for this purpose.

Also the doc and set should pick their arguments immidiately, is
what I feel, while the then also should ideally pick its argument
immidiately and in this case let should ensure that it is picking
a new refernce (or value) each time through the loop.

Need to look into these aspects of JS a bit.